Ref integri

See merge request cvmls/geosoft-1-2023/projektgruppe_01!17

+++ routes/api/DB_api.js
@@ -201,10 +201,34 @@ ROUTER.post('/add_station', checkSchema(GEOJSON_ADD_SCHEMA, ['body']),
   
 });
 
-ROUTER.post('/delete_station', function(req, res) {
+ROUTER.post('/delete_station', async function(req, res) {
+  // To preserve referential integrity, the current tours have to be checked
+  let tours = await fetch("http://localhost:3000/api/tours");
+  tours = await tours.json();
   const ID = req.body.id;
-  delete_item(ID, station_collection);
-  res.send();
+  // All tours where the station is one part get stored, because the user should be explicit informed about them
+  let tours_with_this_station = [];
+  tours.forEach(function(tour) {
+    tour.stations.forEach(function({_id}) {
+      if (_id == ID) {
+        tours_with_this_station.push(tour);
+      }
+    });
+  });
+  // If the station isnt part of any tour it can simply deleted
+  if (tours_with_this_station.length < 1) {
+    delete_item(ID, station_collection);
+    res.json({
+      message : "Alles ok."
+    });
+  }
+  // Else the client gets Error-Message and the tours which contains this station
+  else {
+    res.json({
+      message : "referentielle Integrität gefährdet",
+      tours_with_this_station : tours_with_this_station
+    });
+  }
 });
